File layout target - refactor
