+++ dimos/perception/fiducial/marker_transformer.py
@@ -12,14 +12,15 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-"""ArUco / AprilTag detection as a memory2 transformer.
+"""ArUco / AprilTag detection as memory2 transforms.
 
-Wraps the pure helpers in :mod:`dimos.perception.fiducial.marker_pose`
+Wraps :func:`dimos.perception.fiducial.marker_detect.detect_markers_in_image`
 and emits one :class:`Detection3DMarker` observation per detected marker, with
-``.pose`` composed into world frame from the upstream observation's
-camera-in-world pose. The companion module :class:`MarkerTfModule` remains
-the right choice for live TF publication; this transformer is for offline /
-mem2-stream composition.
+``.pose`` composed into world frame from the upstream observation's camera pose.
+This module also keeps marker smoothing helpers and ``MarkersPerFrame``, which
+collapses marker fan-out back into one ``Detection3DArray`` per source image.
+The companion module :class:`MarkerTfModule` remains the right choice for live
+TF publication.
 
 Skips frames where the upstream observation has no ``.pose`` (debug log):
 without a camera-in-world pose, we can't honor the "always world-frame"
@@ -44,13 +45,12 @@ from dimos.msgs.sensor_msgs.Image import Image
 from dimos.msgs.vision_msgs.Detection3DArray import Detection3DArray
 from dimos.perception.detection.type.detection3d.imageDetections3D import ImageDetections3D
 from dimos.perception.detection.type.detection3d.marker import Detection3DMarker
+from dimos.perception.fiducial.marker_detect import (
+    detect_markers_in_image as _detect_markers_in_image,
+)
 from dimos.perception.fiducial.marker_pose import (
     camera_info_to_cv_matrices,
     create_aruco_detector,
-    estimate_marker_pose,
-    marker_corners_to_bbox,
-    marker_reprojection_error,
-    rvec_tvec_to_transform,
 )
 from dimos.types.timestamped import TimestampedBufferCollection
 from dimos.utils.logging_config import setup_logger
@@ -128,110 +128,6 @@ def _average_marker_pose(
         Vector3(cx, cy, cz),
         Quaternion(qsx / norm, qsy / norm, qsz / norm, qsw / norm),
     )
-
-
-def detect_markers_in_image(
-    image: Image,
-    *,
-    camera_info: CameraInfo,
-    world_T_optical: Transform,
-    marker_length_m: float,
-    aruco_dictionary: str,
-    world_frame: str = "world",
-    detector: Any | None = None,
-    camera_matrix: np.ndarray | None = None,
-    dist_coeffs: np.ndarray | None = None,
-) -> list[Detection3DMarker]:
-    """Detect markers in one image and return rich world-frame 3D detections.
-    """
-    if marker_length_m <= 0:
-        raise ValueError(f"marker_length_m must be > 0, got {marker_length_m}")
-    if (
-        camera_info.width
-        and camera_info.height
-        and (image.width != camera_info.width or image.height != camera_info.height)
-    ):
-        return []
-
-    if detector is None:
-        detector = create_aruco_detector(aruco_dictionary)
-    if (camera_matrix is None) != (dist_coeffs is None):
-        raise ValueError("camera_matrix and dist_coeffs must be provided together")
-    if camera_matrix is None or dist_coeffs is None:
-        camera_matrix, dist_coeffs = camera_info_to_cv_matrices(camera_info)
-
-    gray = image.to_grayscale().as_numpy()
-    corners, ids, _ = detector.detectMarkers(gray)
-    if ids is None or len(ids) == 0:
-        return []
-
-    optical_frame = world_T_optical.child_frame_id or "optical"
-    t_world_optical = Transform(
-        translation=world_T_optical.translation,
-        rotation=world_T_optical.rotation,
-        frame_id=world_frame,
-        child_frame_id=optical_frame,
-        ts=image.ts,
-    )
-    marker_size = Vector3(marker_length_m, marker_length_m, 0.0)
-    detections: list[Detection3DMarker] = []
-
-    for corner_set, mid_arr in zip(corners, ids, strict=True):
-        mid = int(mid_arr[0])
-        pose = estimate_marker_pose(
-            corner_set,
-            marker_length_m,
-            camera_matrix,
-            dist_coeffs,
-            distortion_model=camera_info.distortion_model,
-        )
-        if pose is None:
-            continue
-
-        rvec, tvec = pose
-        t_optical_marker = rvec_tvec_to_transform(
-            rvec,
-            tvec,
-            frame_id=optical_frame,
-            child_frame_id=f"marker_{mid}",
-            ts=image.ts,
-        )
-        t_world_marker = t_world_optical + t_optical_marker
-
-        corners_2d = corner_set.reshape(4, 2).astype(np.float32)
-        bbox = marker_corners_to_bbox(corners_2d)
-        reprojection_error = marker_reprojection_error(
-            corners_2d,
-            marker_length_m,
-            camera_matrix,
-            dist_coeffs,
-            rvec,
-            tvec,
-            distortion_model=camera_info.distortion_model,
-        )
-
-        detections.append(
-            Detection3DMarker(
-                bbox=bbox,
-                track_id=-1,
-                class_id=mid,
-                confidence=1.0,
-                name="",  # set in Detection3DMarker.__post_init__ to marker_label
-                ts=image.ts,
-                image=image,
-                center=t_world_marker.translation,
-                size=marker_size,
-                transform=t_world_optical,
-                frame_id=world_frame,
-                orientation=t_world_marker.rotation,
-                marker_id=mid,
-                corners_px=corners_2d,
-                dictionary=aruco_dictionary,
-                reprojection_error=reprojection_error,
-            )
-        )
-
-    return detections
 
 
 class DetectMarkers(Transformer[Image, Detection3DMarker]):
@@ -324,7 +220,7 @@ class DetectMarkers(Transformer[Image, Detection3DMarker]):
                 ts=obs.ts,
             )
 
-            detections = detect_markers_in_image(
+            detections = _detect_markers_in_image(
                 image,
                 camera_info=info,
                 world_T_optical=t_world_optical,
